For the record: DirectX nieszczegolnie lubi double, trzeba przekazywac jako 2 x uint; radze raczej korzystac z floatow przy shaderach

+++ Hikari Engine/include/Objects/Box.hpp
@@ -16,7 +16,7 @@ namespace Hikari
 		class Box : public Object
 		{
 			public:
-				Box(const Vector3D& center, double height, double width, double depth);
+				Box(const Vector3D& center, float height, float width, float depth);
 
 				~Box();
 		};
