+++ Hikari Engine/src/Objects/Triangle.cpp
@@ -33,6 +33,6 @@ Hikari::Objects::Triangle::~Triangle()
 	m_pIndices = nullptr;
 }
 
-Hikari::Objects::EquilateralTriangle::EquilateralTriangle(const Hikari::Vector3D& triangleCenter, double sideLength) : Triangle(triangleCenter + Hikari::Vector3D(-sideLength/2.0f,-sideLength/3.0f,0.0f), triangleCenter + Hikari::Vector3D(sideLength/2.0f,-sideLength/3,0.0f), triangleCenter + Hikari::Vector3D(0.0f,sideLength*2.0f/3.0f,0.0f))
+Hikari::Objects::EquilateralTriangle::EquilateralTriangle(const Hikari::Vector3D& triangleCenter, float sideLength) : Triangle(triangleCenter + Hikari::Vector3D(-sideLength/2.0f,-sideLength/3.0f,0.0f), triangleCenter + Hikari::Vector3D(sideLength/2.0f,-sideLength/3,0.0f), triangleCenter + Hikari::Vector3D(0.0f,sideLength*2.0f/3.0f,0.0f))
 {
 }
